Plotting refactoring, movements, fix usages
