+++ examples/minimal_run.py
@@ -7,6 +7,7 @@ from core.agent.static import BrownianAgent
 from core.agent.gradient import PhysarumAgent
 from core.agent.base import Agent
 from core.env import Env, Dynamics
+from core.plotting import InteractivePlotter
 from core.utils import setup_logging
 
 
@@ -14,6 +15,7 @@ def run_minimal(agent: Agent, agent_ratio=0.1, field_size=(256, 256), iters=1000
     # Setup the environment
     dynamics = Dynamics(init_agent_ratio=agent_ratio)
     env = Env(field_size, dynamics)
+    plotter = InteractivePlotter.get(env, agent)
 
     total_reward = 0
     obs = env._get_current_obs
@@ -24,7 +26,7 @@ def run_minimal(agent: Agent, agent_ratio=0.1, field_size=(256, 256), iters=1000
         total_reward += reward
         # Visualisation & logging
         pbar.set_postfix(total_reward=np.round(total_reward, 3))
-        env.render(show=True)
+        plotter.draw()
 
 
 if __name__ == '__main__':
