+++ examples/simple_agents.py
@@ -1,5 +1,4 @@
 import numpy as np
-from matplotlib import pyplot as plt
 from tqdm import tqdm, trange
 
 from core.agent.static import ConstAgent, BrownianAgent
@@ -8,6 +7,7 @@ from core.agent.base import Agent
 from core.base_types import ActType
 from core.data_init import WaveSequence
 from core.env import Env, Dynamics
+from core.plotting import InteractivePlotter
 from core.utils import setup_logging
 
 
@@ -28,19 +28,18 @@ def _manual_step(env: Env, action: ActType):
     return env._get_current_obs, 0, False, False, {}
 
 
-def run_agent(env: Env, agent: Agent, iters=1000, show_each=1):
+def run_agent(env: Env, agent: Agent, iters=1000):
     total_reward = 0
     obs = env._get_current_obs
+    plotter = InteractivePlotter.get(env, agent)
 
     for i in (pbar := trange(iters)):
         action = agent.forward(obs)
         obs, reward, _, _, stats = env.step(action)
         total_reward += reward
 
-        if show_each > 0 and i % show_each == 0:
-            pbar.set_postfix(total_reward=np.round(total_reward, 3), **stats)
-            env.render()
-            plt.show()
+        pbar.set_postfix(total_reward=np.round(total_reward, 3), **stats)
+        plotter.draw()
 
 
 def try_const_agent(**kwargs):
